Add files via upload
The Word document contains the report of analysis results, visualizations, and recommendations for the case study.
The .txt file contains a Google Drive link to my folder that contains the Excel spreadsheet files used in the analysis. Due to GitHub's upload limit, I cannot directly upload these files to GitHub.